+++ CHANGELOG.md
@@ -4,8 +4,23 @@
 
 ### Changed
 
+- Update Bitmovin's native Android SDK version to `3.145.0+jason`
+
+## [1.11.0] - 2026-03-05
+### Added
+- Support for streaming ads with the Bitmovin Advertising Module (BAM) via `AdSourceType.BITMOVIN` in `AdvertisingConfig` ad schedules. Note: on iOS, VMAP is currently not supported with BAM.
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.144.0+jason`
+- Update Expo SDK version to `53.0.27`
+
+## [1.10.0] - 2026-02-27
+
+### Changed
+
 - Update Bitmovin's native iOS SDK version to `3.108.1`
-- Update Bitmovin's native Android SDK version to `3.142.0+jason`
+- Update Bitmovin's native Android SDK version to `3.143.0+jason`
 - Update Expo SDK version to `53.0.26`
 
 ### Fixed